avoiding roundoff errors in ft_spike_maketrials.m
git-svn-id: svn+ssh://svn.fcdonders.nl/home/svnroot/fieldtrip/trunk@7622 0cf7c7f0-3615-4144-b4e6-68da3bce3cd0

+++ contrib/spike/ft_spike_maketrials.m
@@ -206,7 +206,7 @@ elseif strcmp(cfg.trlunit,'samples')
 end
 
 if size(cfg.trl,2) > 3
-    spike.trialinfo        = cfg.trl(:,4:end);
+    spike.trialinfo        = double(cfg.trl(:,4:end));
 end
 
 % do the general cleanup and bookkeeping at the end of the function
